Refactored, values updating split into distinct threads for events and bg values. - Refactored to separate VoiceAttack plugin execution methods from VoiceAttack variable setting methods. - Split background variable setting to a separate thread to prevent delays from the plugin in reacting to events. - Revised the method for iterating object data and writing values to VoiceAttack. It now writes more properties than it did previously (auto-generated docs do not fully describe the currently written VoiceAttack variables that are generated for object data even though they *are* present).
